backend/requirements.txt
Flask
Flask-CORS
Flask-RESTful
Jinja2
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
